backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "cryptography>=46.0.6",
    "docx2txt>=0.9",
    "fastapi>=0.135.3",
    "langchain>=1.2.15",
    "langchain-community>=0.4.1",
    "langchain-openai>=1.1.12",
    "langgraph>=1.1.6",
    "load-dotenv>=0.1.0",
    "logging>=0.4.9.6",
    "mcp>=1.27.0",
    "passlib[bcrypt]>=1.7.4",
    "pymilvus[milvus-lite]>=2.6.12",
    "pymysql>=1.1.2",
    "pypdf>=6.10.2",
    "pymupdf>=1.26.5",
    "python-jose>=3.5.0",
    "pytz>=2026.1.post1",
    "requests>=2.33.1",
    "rapidocr-onnxruntime>=1.4.4",
    "sentence-transformers>=5.4.0",
    "sqlalchemy>=2.0.49",
    "uvicorn>=0.42.0",
    "pytesseract>=0.3.13",
]
